requirements.txt
Flask==0.10.1
Flask-WTF==0.9.3
Jinja2==2.7.1
SQLAlchemy==0.8.3
Werkzeug==0.9.4
WTForms==1.0.5
itsdangerous==0.23
requests==1.2.3
# Currently this is https://github.com/dotcloud/docker-py/archive/35d3a3c977ba3a08ddf61bb80a6c8bccab187cc1.zip
docker-py==0.2.2
mock==1.0.1
websocket-client==0.11.0
sh==1.09
python-dateutil==2.2
livereload==1.0.1
uWSGI==1.9.20
six==1.3.0
structlog==0.4.0

#docopt==0.6.1
#mock==1.0.1
#websocket-client==0.11.0
